backend/requirements.txt

fastapi
uvicorn
openai
gtts
google-cloud-storage
